feat(dfir_lang): add `codegen` feature to avoid heavy deps when only key types are needed
Problem: PR #2810 made `dfir_lang` a non-optional dependency of `dfir_rs` for
metrics (which needs `GraphNodeId`/`GraphSubgraphId` slotmap keys). This pulled
~69 crates (syn, proc-macro2, quote, prettyplease, documented, etc.) into target
compilation even when `meta` is disabled — defeating the trybuild latency
optimization.

Solution: Add a `codegen` feature (default-enabled) to `dfir_lang` that gates
all heavy dependencies. Without it, only `slotmap` + `serde` are needed.

Changes:
- dfir_lang/src/graph_ids.rs: New module with just the 4 slotmap key type
  definitions (GraphNodeId, GraphEdgeId, GraphSubgraphId, GraphLoopId).
- dfir_lang/src/lib.rs: Always expose `graph_ids`; gate `graph`, `diagnostic`,
  `parse`, `pretty_span`, `process_singletons`, `union_find` behind
  `#[cfg(feature = "codegen")]`.
- dfir_lang/src/graph/mod.rs: Re-export key types from `graph_ids` (preserving
  the `dfir_lang::graph::{GraphNodeId, ...}` path for existing users).
- dfir_lang/Cargo.toml: New `codegen` feature (default) gating syn, proc-macro2,
  quote, prettyplease, documented, auto_impl, itertools, serde_json.
- dfir_rs/Cargo.toml: Depend on `dfir_lang` with `default-features = false`;
  `meta` feature enables `dfir_lang/codegen`.
- dfir_rs/src/scheduled/metrics.rs: Import from `dfir_lang::graph_ids` instead
  of `dfir_lang::graph`.
- dfir_lang/src/graph/meta_graph.rs: Fix private field access (`subgraph_id.0`
  → `subgraph_id.data()`).

Result: `dfir_lang` without `codegen` pulls only slotmap+serde (~3 target crates
beyond what dfir_rs already has), vs ~69 crates with full codegen.

PR: #2810
